Backend/requirements.txt
flask
flask-cors
pillow
opencv-python
numpy
svgwrite
requests
imageio
scikit-image
scipy
